Moved IPauseManager from server to shared.
