+++ Robust.Server/GameObjects/ServerEntityManager.cs
@@ -4,7 +4,6 @@ using System.Linq;
 using System.Threading;
 using Prometheus;
 using Robust.Server.Player;
-using Robust.Server.Timing;
 using Robust.Shared;
 using Robust.Shared.Configuration;
 using Robust.Shared.GameObjects;
