+++ Robust.Server/Maps/MapLoader.cs
@@ -12,9 +12,9 @@ using Robust.Shared.GameObjects;
 using System.Globalization;
 using System.Linq;
 using Robust.Server.GameObjects;
-using Robust.Server.Timing;
 using Robust.Shared.ContentPack;
 using Robust.Shared.Prototypes;
+using Robust.Shared.Timing;
 using YamlDotNet.Core;
 
 namespace Robust.Server.Maps
