+++ Robust.Server/ServerIoC.cs
@@ -10,7 +10,6 @@ using Robust.Server.Prototypes;
 using Robust.Server.Reflection;
 using Robust.Server.Scripting;
 using Robust.Server.ServerStatus;
-using Robust.Server.Timing;
 using Robust.Server.ViewVariables;
 using Robust.Shared;
 using Robust.Shared.ContentPack;
@@ -20,6 +19,7 @@ using Robust.Shared.Map;
 using Robust.Shared.Network;
 using Robust.Shared.Prototypes;
 using Robust.Shared.Reflection;
+using Robust.Shared.Timing;
 
 namespace Robust.Server
 {
@@ -41,7 +41,6 @@ namespace Robust.Server
             IoCManager.Register<IEntityNetworkManager, ServerEntityNetworkManager>();
             IoCManager.Register<IServerEntityNetworkManager, ServerEntityNetworkManager>();
             IoCManager.Register<IMapLoader, MapLoader>();
-            IoCManager.Register<IPauseManager, PauseManager>();
             IoCManager.Register<IPlacementManager, PlacementManager>();
             IoCManager.Register<IPlayerManager, PlayerManager>();
             IoCManager.Register<IPrototypeManager, ServerPrototypeManager>();
