--- Robust.Shared/GameObjects/Components/SharedIgnorePauseComponent.cs
@@ -1,7 +1,0 @@
-namespace Robust.Shared.GameObjects
-{
-    public abstract class SharedIgnorePauseComponent : Component
-    {
-        public override string Name => "IgnorePause";
-    }
-}
